276: Remove the configmap from the chart values
The config map will be available in sbat namespace and is currently defined
in the gateway deployment - this can be split out at a later date into it's own
repo. For now, we just need to prevent the config map install being done
by the iam initializer helm install, as this was only needed when the
initializer was installed into a CDK namespace.

Align and rationalise environment variable names

Issue: https://github.com/securebankingaccesstoolkit/securebankingaccesstoolkit/issues/276