621: Removing the unique constraint from apiClient.ssa field
There is no requirement in the spec that an ssa needs to be unique, multiple apiClients should be able to reference the same ssa.

In our current environments this is the case, as we have a bug where the ssa field is not being set. In order to fix that bug we need to disable the unique constraint.

Also doing some cleanup to make the id, name, oauth2ClientId and ssa required fields.

https://github.com/SecureBankingAccessToolkit/SecureBankingAccessToolkit/issues/621