694: Configure GoogleSecretStore for CDK environments only.
FIDC environments come pre-configured with Google Secrets Manager config. The settings of the manager may not be editted.

For FIDC we still want to be able to map the secrets to the store, so the name of the store in the config must match the name of the pre-configured store (which is "ESV")

Issue: https://github.com/SecureBankingAccessToolkit/SecureBankingAccessToolkit/issues/694