779: Configuring AM Remote Consent from environment variables
Added logic to create a JWKS from an RSA public key. Configuring this JWKS in AM for it to verify JWTs signed by the RCS.

Mounted the public signing key from the rcs-signing secret.
Configured environment variables for the signing keyId and jwt issuer, so that AM and RCS are configured consistently.

Issue: https://github.com/SecureBankingAccessToolkit/SecureBankingAccessToolkit/issues/779