1529: Rename ApiClient.id field to softwareId
The ApiClient.id field is confusing, it is not the unique id of an ApiClient object, the _id field should be used for this purpose.

The field actually represents the softwareId (which is not unique, multiple clients may be registered for the same piece of software).

In this change we are "renaming" the id field in a non breaking fashion by doing the following:
- Adding new field: softwareId which is optional
- Making the existing id field optional.
- Adding an onRead script which defaults the softwareId field to the id value

Using the script prevents this from being a breaking change by providing a default value for the softwareId.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1529